+++ contrib/opencensus-ext-azure/examples/logs/properties.py
@@ -32,5 +32,3 @@ try:
     result = 1 / 0  # generate a ZeroDivisionError
 except Exception:
     logger.exception('Captured an exception.', extra=properties)
-
-input("...")