+++ contrib/opencensus-ext-azure/opencensus/ext/azure/metrics_exporter/__init__.py
@@ -14,6 +14,7 @@
 
 import atexit
 import logging
+import os
 
 from opencensus.common import utils as common_utils
 from opencensus.ext.azure.common import Options, utils
@@ -39,8 +40,9 @@ logger = logging.getLogger(__name__)
 class MetricsExporter(TransportMixin, ProcessorMixin):
     """Metrics exporter for Microsoft Azure Monitor."""
 
-    def __init__(self, **options):
+    def __init__(self, is_stats=False, **options):
         self.options = Options(**options)
+        self._is_stats = is_stats
         utils.validate_instrumentation_key(self.options.instrumentation_key)
         if self.options.max_batch_size <= 0:
             raise ValueError('Max batch size must be at least 1.')
@@ -58,6 +60,8 @@ class MetricsExporter(TransportMixin, ProcessorMixin):
             )
         self._atexit_handler = atexit.register(self.shutdown)
         self.exporter_thread = None
+        # For redirects
+        self._consecutive_redirects = 0  # To prevent circular redirects
         super(MetricsExporter, self).__init__()
 
     def export_metrics(self, metrics):
@@ -131,7 +135,10 @@ class MetricsExporter(TransportMixin, ProcessorMixin):
             tags=dict(utils.azure_monitor_context),
             time=timestamp.isoformat(),
         )
-        envelope.name = "Microsoft.ApplicationInsights.Metric"
+        if self._is_stats:
+            envelope.name = "Statsbeat"
+        else:
+            envelope.name = "Microsoft.ApplicationInsights.Metric"
         data = MetricData(
             metrics=[data_point],
             properties=properties
@@ -141,7 +148,8 @@ class MetricsExporter(TransportMixin, ProcessorMixin):
 
     def shutdown(self):
         # Flush the exporter thread
-        if self.exporter_thread:
+        # Do not flush if metrics exporter for stats
+        if self.exporter_thread and not self._is_stats:
             self.exporter_thread.close()
         # Shutsdown storage worker
         if self.storage:
@@ -157,4 +165,8 @@ def new_metrics_exporter(**options):
                                     producers,
                                     exporter,
                                     interval=exporter.options.export_interval)
+    if not os.environ.get("APPLICATIONINSIGHTS_STATSBEAT_DISABLED_ALL"):
+        from opencensus.ext.azure.metrics_exporter import statsbeat_metrics
+        # Stats will track the user's ikey
+        statsbeat_metrics.collect_statsbeat_metrics(exporter.options)
     return exporter