+++ contrib/opencensus-ext-azure/opencensus/ext/azure/trace_exporter/__init__.py
@@ -15,6 +15,7 @@
 import atexit
 import json
 import logging
+import os
 
 from opencensus.common.schedule import QueueExitEvent
 from opencensus.ext.azure.common import Options, utils
@@ -29,6 +30,7 @@ from opencensus.ext.azure.common.protocol import (
 )
 from opencensus.ext.azure.common.storage import LocalFileStorage
 from opencensus.ext.azure.common.transport import TransportMixin
+from opencensus.ext.azure.metrics_exporter import statsbeat_metrics
 from opencensus.trace import attributes_helper
 from opencensus.trace.span import SpanKind
 
@@ -72,6 +74,11 @@ class AzureExporter(BaseExporter, ProcessorMixin, TransportMixin):
         self._telemetry_processors = []
         super(AzureExporter, self).__init__(**options)
         atexit.register(self._stop, self.options.grace_period)
+        # start statsbeat on exporter instantiation
+        if not os.environ.get("APPLICATIONINSIGHTS_STATSBEAT_DISABLED_ALL"):
+            statsbeat_metrics.collect_statsbeat_metrics(self.options)
+        # For redirects
+        self._consecutive_redirects = 0  # To prevent circular redirects
 
     def span_data_to_envelope(self, sd):
         envelope = Envelope(